+++ plugins/product/src/Controllers/Web/DesignIdeaController.php
@@ -192,7 +192,7 @@ class DesignIdeaController extends BasePublicController{
     {
         $lookBookId = get_id_from_url($url);
         $lookBook = $this->lookBookServices->getDetailLookBook($lookBookId);
-        $productWishListIds = $this->wishListRepositories->getArrayIdWishListProductsByCustomer(Auth::guard('customer')->id());
+        $productWishListIds = $this->wishListRepositories->getArrayIdWishListProductsByCustomer((int)Auth::guard('customer')->id());
         AssetManager::addAsset('look-book-design-css', 'frontend/plugins/product/assets/css/look-book-design.css');
         AssetPipeline::requireCss('look-book-design-css');
         AssetManager::addAsset('design-idea-js', 'frontend/plugins/product/assets/js/design-idea.js');
