+++ plugins/product/src/Controllers/Web/ProductCategoryController.php
@@ -50,7 +50,7 @@ class ProductCategoryController extends BasePublicController
         AssetPipeline::requireJs('product-detail-js');
         AssetManager::addAsset('product-detail-css', 'frontend/plugins/product/assets/css/product-detail.css');
         AssetPipeline::requireCss('product-detail-css');
-        $productWishListIds = $this->wishListRepositories->getArrayIdWishListProductsByCustomer(Auth::guard('customer')->id());
+        $productWishListIds = $this->wishListRepositories->getArrayIdWishListProductsByCustomer((int)Auth::guard('customer')->id());
         if (empty($categoryPageInfo['category']->parent_id))
             return view('pages.category.detail', compact('categoryPageInfo', 'productWishListIds'));
         else
@@ -64,7 +64,7 @@ class ProductCategoryController extends BasePublicController
     public function getListSaleProductsOfCategoryPage($url) {
         $productCategoryId = get_id_from_url($url);
         $categoryPageInfo = $this->productServices->getListProductsOfCategoryPage($productCategoryId);
-        $productWishListIds = $this->wishListRepositories->getArrayIdWishListProductsByCustomer(Auth::guard('customer')->id());
+        $productWishListIds = $this->wishListRepositories->getArrayIdWishListProductsByCustomer((int)Auth::guard('customer')->id());
         AssetManager::addAsset('product-detail-js', 'frontend/plugins/product/assets/js/product-detail.js');
         AssetPipeline::requireJs('product-detail-js');
         AssetManager::addAsset('product-detail-css', 'frontend/plugins/product/assets/css/product-detail.css');
